refactor: getCardInfo을 커스텀 훅으로 분리
useCardPreview 커스텀 훅 생성함
